File reading
Setup a midiFiles folder that isnt reflected in the repo. The
filecontroller reads from that directory, listing all files held there.
Going to implement selection, playback, and midifile write.

+++ midiParser/MidiGate.java
@@ -29,7 +29,7 @@ public class MidiGate {
 		
 		for( jm.music.data.Note n : notes){
 			
-			Note newNote = new Note(n.getPitch(), (int)(n.getDuration() * beatInMS) , n.getDynamic());
+			Note newNote = new Note(n.getPitch(), (long)(n.getDuration() * beatInMS) , n.getDynamic());
 			trumpNotes.add(newNote);
 			
 		}
